pi_agent/requirements.txt
requests
pydantic-settings
rpi-lgpio; platform_system == "Linux"
